refactor(tests): group test helpers under tests/*/helpers/ subdirs
Move loose non-test Python modules into conventional helpers/ subdirs
so the test-naming hook can exclude a single pattern instead of
enumerating each file.

Moves:
- tests/e2e/mcp_test_helpers.py -> tests/e2e/helpers/mcp_test_helpers.py
- tests/protocol_compliance/_helpers.py ->
  tests/protocol_compliance/helpers/compliance.py
- tests/protocol_compliance/_drift.py ->
  tests/protocol_compliance/helpers/drift.py

Updates all import sites across tests/e2e, tests/e2e_rust, and
tests/protocol_compliance.

Config cleanup:
- name-tests-test now excludes tests/*/(helpers|fixtures|targets|pages)/
  as a single verbose-regex alternative instead of file-by-file.
- Top-level pre-commit exclude and detect-secrets excludes converted
  to verbose-regex with inline comments.
- Dropped non-existent excludes (scripts/sign_image.sh, scripts/zap,
  sonar-project.properties) and the redundant .secrets.baseline
  entry (detect-secrets skips its own baseline automatically).
- detect-secrets hook in pre-commit now carries its own exclude list,
  kept in sync with the Makefile invocation.
