+++ tests/e2e/helpers/mcp_test_helpers.py
