+++ tests/e2e/test_mcp_plugin_parity.py
@@ -24,7 +24,7 @@ import pytest
 
 from mcpgateway.utils.create_jwt_token import _create_jwt_token
 
-from tests.e2e.mcp_test_helpers import BASE_URL, skip_no_gateway
+from tests.e2e.helpers.mcp_test_helpers import BASE_URL, skip_no_gateway
 
 MCP_PROTOCOL_VERSION = "2025-11-25"
 PLUGIN_PARITY_PREFIX = "mcp-plugin-parity"
@@ -39,10 +39,7 @@ pytestmark = [
     skip_no_gateway,
     pytest.mark.skipif(
         not EXPECTED_RUNTIME,
-        reason=(
-            "requires the dedicated plugin parity stack; run via "
-            "MCP_PLUGIN_PARITY_EXPECTED_RUNTIME=<python|rust> make test-mcp-plugin-parity"
-        ),
+        reason=("requires the dedicated plugin parity stack; run via " "MCP_PLUGIN_PARITY_EXPECTED_RUNTIME=<python|rust> make test-mcp-plugin-parity"),
     ),
 ]
 
@@ -120,9 +117,7 @@ def _request_json(
         Parsed JSON response body, or `None` for an empty body.
     """
     response = client.request(method, path, **kwargs)
-    assert response.status_code in expected, (
-        f"{method} {path} expected {expected}, got {response.status_code}: {response.text}"
-    )
+    assert response.status_code in expected, f"{method} {path} expected {expected}, got {response.status_code}: {response.text}"
     return response.json() if response.content else None
