+++ tests/protocol_compliance/test_notifications.py
@@ -7,7 +7,7 @@ import json
 
 import pytest
 
-from ._helpers import resolve_tool, xfail_on
+from .helpers.compliance import resolve_tool, xfail_on
 
 pytestmark = [pytest.mark.protocol_compliance, pytest.mark.mcp_utilities]
