+++ tests/protocol_compliance/test_resources.py
@@ -7,7 +7,7 @@ import json
 import pytest
 from fastmcp.client import Client
 
-from ._helpers import xfail_on
+from .helpers.compliance import xfail_on
 
 pytestmark = [pytest.mark.protocol_compliance, pytest.mark.mcp_server_features]
