+++ tests/protocol_compliance/test_tools.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pytest
 from fastmcp.client import Client
 
-from ._helpers import resolve_tool, xfail_on
+from .helpers.compliance import resolve_tool, xfail_on
 
 pytestmark = [pytest.mark.protocol_compliance, pytest.mark.mcp_server_features]
